src/infra/http/axios-http-client.ts
import axios, { AxiosResponse } from 'axios';
import {
  HttpClient,
  HttpRequest,
  HttpResponse,
} from '../../data/prootocols/http/http-client';

export default class AxiosHttpClient implements HttpClient {
  async request(data: HttpRequest): Promise<HttpResponse> {
    let axiosResponse: AxiosResponse;
    try {
      axiosResponse = await axios.request({
        url: data.url,
        method: data.method,
        data: data.body,
        headers: data.headers,
        params: data.params,
      });
    } catch (error: any) {
      axiosResponse = error.response;
    }

    return {
      statusCode: axiosResponse.status,
      body: axiosResponse.data,
    };
  }
}
